Fix some problems
1. When creating or editing a document - incorrectly indicates the
default value, such as whether to allow to place the document in the map
by default YES, in the form specified - NO.

2. If you do not press the TAB seo in shape while editing the data form
does not go away. And the idea that must be replaced by default - does
not work. As a result, we have a database in the properties of the
component, if you use the default publication = 1, we put 0.

3. duplicates URL